Misc cleanups and refactoring
