+++ cardano_node_tests/tests/test_configuration.py
@@ -10,6 +10,7 @@ import pytest
 from _pytest.tmpdir import TempdirFactory
 from cardano_clusterlib import clusterlib
 
+from cardano_node_tests.tests import common
 from cardano_node_tests.utils import cluster_management
 from cardano_node_tests.utils import cluster_nodes
 from cardano_node_tests.utils import helpers
@@ -40,7 +41,7 @@ pytestmark = pytest.mark.usefixtures("temp_dir")
 @pytest.fixture(scope="module")
 def epoch_length_start_cluster(tmp_path_factory: TempdirFactory) -> Path:
     """Update *epochLength* to 1200."""
-    pytest_globaltemp = helpers.get_pytest_globaltemp(tmp_path_factory)
+    pytest_globaltemp = common.get_pytest_globaltemp(tmp_path_factory)
 
     # need to lock because this same fixture can run on several workers in parallel
     with helpers.FileLockIfXdist(f"{pytest_globaltemp}/startup_files_epoch_1200.lock"):
@@ -69,7 +70,7 @@ def epoch_length_start_cluster(tmp_path_factory: TempdirFactory) -> Path:
 @pytest.fixture(scope="module")
 def slot_length_start_cluster(tmp_path_factory: TempdirFactory) -> Path:
     """Update *slotLength* to 0.3."""
-    pytest_globaltemp = helpers.get_pytest_globaltemp(tmp_path_factory)
+    pytest_globaltemp = common.get_pytest_globaltemp(tmp_path_factory)
 
     # need to lock because this same fixture can run on several workers in parallel
     with helpers.FileLockIfXdist(f"{pytest_globaltemp}/startup_files_slot_03.lock"):
