+++ cardano_node_tests/tests/test_delegation.py
@@ -9,11 +9,11 @@ import pytest
 from _pytest.tmpdir import TempdirFactory
 from cardano_clusterlib import clusterlib
 
+from cardano_node_tests.tests import common
 from cardano_node_tests.tests import delegation
 from cardano_node_tests.utils import cluster_management
 from cardano_node_tests.utils import cluster_nodes
 from cardano_node_tests.utils import clusterlib_utils
-from cardano_node_tests.utils import constants
 from cardano_node_tests.utils import dbsync_utils
 from cardano_node_tests.utils import helpers
 
@@ -103,9 +103,7 @@ class TestDelegateAddr:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
@@ -157,9 +155,7 @@ class TestDelegateAddr:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
@@ -474,9 +470,7 @@ class TestDelegateAddr:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
@@ -560,9 +554,7 @@ class TestDelegateAddr:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
@@ -822,9 +814,7 @@ class TestNegative:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
@@ -889,9 +879,7 @@ class TestNegative:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
@@ -988,9 +976,7 @@ class TestNegative:
             False,
             pytest.param(
                 True,
-                marks=pytest.mark.skipif(
-                    not constants.BUILD_USABLE, reason=constants.BUILD_SKIP_MSG
-                ),
+                marks=pytest.mark.skipif(not common.BUILD_USABLE, reason=common.BUILD_SKIP_MSG),
             ),
         ),
         ids=("build_raw", "build"),
