+++ cardano_node_tests/utils/configuration.py
@@ -3,6 +3,17 @@ import os
 from pathlib import Path
 
 
+LAUNCH_PATH = Path(os.getcwd())
+
+NETWORK_MAGIC_LOCAL = 42
+DBSYNC_DB = "dbsync"
+
+TESTNET_POOL_IDS = (
+    "pool18yslg3q320jex6gsmetukxvzm7a20qd90wsll9anlkrfua38flr",
+    "pool15sfcpy4tps5073gmra0e6tm2dgtrn004yr437qmeh44sgjlg2ex",
+    "pool1csh8x6227uphxz67nr8qhmd8c7nsyct2ptn7t0yjkhqu7neauwu",
+)
+
 # resolve CARDANO_NODE_SOCKET_PATH
 os.environ["CARDANO_NODE_SOCKET_PATH"] = str(
     Path(os.environ["CARDANO_NODE_SOCKET_PATH"]).expanduser().resolve()
