+++ cardano_node_tests/utils/dbsync_conn.py
@@ -6,7 +6,7 @@ from typing import Optional
 import psycopg2
 
 from cardano_node_tests.utils import cluster_nodes
-from cardano_node_tests.utils import constants
+from cardano_node_tests.utils import configuration
 from cardano_node_tests.utils import helpers
 
 LOGGER = logging.getLogger(__name__)
@@ -21,7 +21,7 @@ class DBSyncCache:
 def _conn(instance_num: int) -> psycopg2.extensions.connection:
     # Call `psycopg2.connect` with an empty string so it uses PG* env variables.
     # Temporarily set PGDATABASE env var to the database corresponding to `instance_num`.
-    with helpers.environ({"PGDATABASE": f"{constants.DBSYNC_DB}{instance_num}"}):
+    with helpers.environ({"PGDATABASE": f"{configuration.DBSYNC_DB}{instance_num}"}):
         conn = psycopg2.connect("")
     DBSyncCache.conns[instance_num] = conn
     return conn
@@ -31,13 +31,13 @@ def _close(instance_num: int, conn: Optional[psycopg2.extensions.connection]) ->
     if conn is None or conn.closed == 1:
         return
 
-    LOGGER.info(f"Closing connection to db-sync database {constants.DBSYNC_DB}{instance_num}.")
+    LOGGER.info(f"Closing connection to db-sync database {configuration.DBSYNC_DB}{instance_num}.")
     try:
         conn.close()
     except psycopg2.Error as err:
         LOGGER.warning(
             "Unable to close connection to db-sync database "
-            f"{constants.DBSYNC_DB}{instance_num}: {err}"
+            f"{configuration.DBSYNC_DB}{instance_num}: {err}"
         )
